ai-service/pyproject.toml
[project]
name = "ai-service"
version = "0.0.1"
description = "YetBota AI service: content dedup, RAG assistant, profanity screening."
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "aio-pika>=9.5,<10.0",
    "weaviate-client>=4.9,<5.0",
    "neo4j>=5.25,<6.0",
    "google-genai>=0.3",
    "structlog>=24.4",
    "tenacity>=9.0",
    "grpcio>=1.68",
    "grpcio-reflection>=1.68",
    "protobuf>=5.28",
    "prometheus-client>=0.21",
    "jinja2>=3.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ASYNC", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]
plugins = ["pydantic.mypy"]
mypy_path = "src:../common/proto/generated/python"
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["ai.*", "google.genai.*", "weaviate.*", "aio_pika.*", "grpc.*"]
ignore_missing_imports = true

[tool.importlinter]
root_packages = ["domain", "application", "infrastructure", "interfaces"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "domain has no outbound deps"
type = "forbidden"
source_modules = ["domain"]
forbidden_modules = [
    "application",
    "infrastructure",
    "interfaces",
    "fastapi",
    "uvicorn",
    "grpc",
    "aio_pika",
    "weaviate",
    "google.genai",
]

[[tool.importlinter.contracts]]
name = "application depends only on domain"
type = "forbidden"
source_modules = ["application"]
forbidden_modules = [
    "infrastructure",
    "interfaces",
    "fastapi",
    "uvicorn",
    "grpc",
    "aio_pika",
    "weaviate",
    "google.genai",
]
